# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AaabPashi Web API
  description: |
    AaabPashi is a comprehensive water management and agricultural monitoring platform API.

    The API provides authentication via phone number and OTP, field management, geospatial 
    data processing, water release scheduling, and multilingual support.

    ## Authentication

    This API supports two authentication methods:

    ### 1. JWT-based Authentication (Primary)
    JWT-based authentication with phone number and OTP verification. 
    Authentication tokens are stored in HTTP-only cookies for security.

    ### 2. API Key Authentication (Alternative)
    API key authentication for programmatic access, third-party integrations, 
    and administrative operations. API keys can be provided in headers:
    - `Authorization: Bearer <api-key>`
    - `X-API-Key: <api-key>`

    ### Authentication Priority
    1. API key authentication is checked first
    2. If API key fails, JWT authentication is attempted
    3. If both fail, request is rejected with 401 Unauthorized

    ## Rate Limiting

    OTP endpoints are rate-limited to one request per phone number every 5 minutes.

    ## External Services

    The API integrates with:
    - Veevo SMS API for OTP delivery
    - Google Translate API for multilingual support
    - Open-Meteo API for weather data
    - Google Earth Engine API for satellite imagery
    - Custom Python APIs for geospatial processing

    ## Multi-Platform Sync

...
  version: 1.0.0
...
servers:
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /{locale}/api/auth/verifyotp-register:
    post:
      tags:
        - Authentication
      summary: Verify OTP for registration
      description: Verify OTP and complete user registration
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
            enum: [en, ur]
          description: Language locale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPRegisterRequest"
      responses:
        "200":
          description: User registered successfully
          headers:
            Set-Cookie:
              description: Authentication cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid OTP or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/getfield:
    post:
      tags:
        - Field Management
      summary: Get user fields
      description: Retrieve all fields for a specific user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetFieldRequest"
      responses:
        "200":
          description: Fields retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetFieldResponse"
        "400":
          description: Missing user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/getGEE:
    post:
      tags:
        - Geospatial
      summary: Get Google Earth Engine data
      description: Get Google Earth Engine satellite data for a polygon
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetGEERequest"
      responses:
        "200":
          description: GEE data retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GEEData"
        "400":
          description: Polygon required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to contact GEE API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/sync/create-user:
    post:
      tags:
        - User Sync
      summary: Sync user create (internal)
      description: Create or upsert a user in the local database (API key protected, internal use only)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUser"
      responses:
        "200":
          description: User synced (created/updated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSuccessResponse"
        "400":
          description: Missing user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sync/update-user:
    post:
      tags:
        - User Sync
      summary: Sync user update (internal)
      description: Update a user in the local database (API key protected, internal use only)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUser"
      responses:
        "200":
          description: User synced (updated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSuccessResponse"
        "400":
          description: Missing user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sync/delete-user:
    post:
      tags:
        - User Sync
      summary: Sync user delete (internal)
      description: Delete a user from the local database (API key protected, internal use only)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserDelete"
      responses:
        "200":
          description: User synced (deleted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSuccessResponse"
        "400":
          description: Missing user identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only cookie
  schemas:
...
    VerifyOTPRegisterRequest:
      type: object
      required:
        - phoneNumber
        - otp
        - name
        - city
        - division
        - receiverNetwork
        - farmsize
        - role
        - country
      properties:
        phoneNumber:
          type: string
          example: "+923001234567"
        otp:
          type: string
          example: "1234"
        name:
          type: string
          example: "John Doe"
        city:
          type: string
          example: "Lahore"
        division:
          type: string
          example: "Kasur"
        receiverNetwork:
          type: string
          example: "Jazz"
        farmsize:
          type: string
          example: "5-10 acres"
        role:
          type: string
          example: "Farmer"
        country:
          type: string
          example: "Pakistan"
...
    GetFieldRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          example: "64a1b2c3d4e5f6789012345"
    GetFieldResponse:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/Field"
...
    GetGEERequest:
      type: object
      required:
        - polygon
      properties:
        polygon:
          $ref: "#/components/schemas/GeoJSONPolygon"
    GEEData:
      type: object
      properties:
        ndvi:
          type: number
          format: float
          example: 0.75
        precipitation:
          type: number
          format: float
          example: 25.5
        temperature:
          type: number
          format: float
          example: 28.3
        soil_moisture:
          type: number
          format: float
          example: 0.65
...
    Location:
      type: object
      required:
        - lat
        - lng
      properties:
        lat:
          type: number
          format: float
          minimum: -90
          maximum: 90
          example: 31.5204
        lng:
          type: number
          format: float
          minimum: -180
          maximum: 180
          example: 74.3587
...
    GeoJSONPolygon:
      type: object
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum: [Polygon]
          example: "Polygon"
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
                format: float
              minItems: 2
              maxItems: 2
          example:
            [
              [
                [74.3587, 31.5204],
                [74.3590, 31.5204],
                [74.3590, 31.5207],
                [74.3587, 31.5207],
                [74.3587, 31.5204],
              ],
            ]
    Field:
      type: object
      properties:
        _id:
          type: string
          example: "64a1b2c3d4e5f6789012346"
        fieldName:
          type: string
          example: "North Field"
        cropTypes:
          type: array
          items:
            type: string
          example: ["Wheat", "Cotton"]
        soilType:
          type: string
          example: "Clay"
        location:
          $ref: "#/components/schemas/Location"
        polygon:
          $ref: "#/components/schemas/GeoJSONPolygon"
        userId:
          type: string
          example: "64a1b2c3d4e5f6789012345"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"

    # API Key Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        error:
          type: string
          example: "Detailed error information"
    SyncUser:
      type: object
      required:
        - name
        - phone
        - city
        - division
        - role
      properties:
        name:
          type: string
          example: John Doe
        phone:
          type: string
          example: "+923001234567"
        city:
          type: string
          example: Lahore
        division:
          type: string
          example: Kasur
        role:
          type: string
          example: Farmer
        # ... other user fields ...
    SyncSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User synced (created/updated)
    SyncUserDelete:
      type: object
      properties:
        phone:
          type: string
          example: "+923001234567"
        _id:
          type: string
          example: "64a1b2c3d4e5f6789012345"